2. legacy test projects/spring-data-jpa-practice/src/main/java/practice/spring/data/jpa/doing/v2/UserRepository.java
package practice.spring.data.jpa.doing.v2;

import java.util.List;
import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;
import org.springframework.stereotype.Repository;

//@Repository
class UserRepository {

    @PersistenceContext
    private EntityManager em;

    public UserRepository(EntityManager em) {
        this.em = em;
    }

    public User save(User user) {
        em.persist(user);
        return user;
    }

    public List<User> findAll() {
        return em.createQuery("select u from User u", User.class)
                .getResultList();
    }
}
